# CUDA: Improve flash decoding kernel GPU occupancy for BS=1 case