This PR adds the following optimizations to the CUDA flash decoding code:
- Find out active blocks per SM using `cudaOccupancyMaxActiveBlocksPerMultiprocessor` API. Use this value to determine the optimal `parallel_blocks` value.
- Prefer vector flash attention kernels over MMA kernel for BS=1 as with this change vector_mha kernel beats mma based mha kernels.

As flash decoding is dominant only with large seq lengths, e2e perf improvements are seen only with larger seq lengths. With 16000 seq length, we see an improvement of up to **14%** in gen phase throughput. The gains are expected to continue to scale with larger seq lengths. Details available below.

Issue: #12182 

<details>
<summary><b>Performance</b></summary>

As this change impacts gen phase throughput for large seq lengths, the benchmarking was done in following way:
- Run `llama-bench.exe -m <model> -p 10,100,1000,2000,4000,8000,16000 -n 0 -pg 10,200 -pg 100,200 -pg 1000,200 -pg 2000,200 -pg 4000,200 -pg 8000,200 -pg 16000,200 -o csv -fa 1`
- This provides `avg_ns` value for a. the prompt phase and b. the prompt + gen phase combined. Subtracting #a) from #b) gives time spent in only gen phase which is shown as `gen_avg_ms` in below table.

I get the following numbers on RTX 4090 on master (prefixed with master) and this pr (prefixed with pr):

| Model                    | n_prompt | n_gen | master_avg_ns | master_gen_avg_ms | pr_avg_ns  | pr_gen_avg_ms | Speed-up |
|--------------------------|----------|-------|---------------|-------------------|------------|---------------|----------|
| llama 3B Q4_K - Medium   |          |       |               |                   |            |               |          |
|                          | 10       | 0     | 8276500       |                   | 8210860    |               |          |
|                          | 100      | 0     | 11219420      |                   | 11776440   |               |          |
|                          | 1000     | 0     | 42959940      |                   | 44420860   |               |          |
|                          | 2000     | 0     | 86611920      |                   | 86546600   |               |          |
|                          | 4000     | 0     | 184733060     |                   | 180640240  |               |          |
|                          | 8000     | 0     | 398559780     |                   | 398832300  |               |          |
|                          | 16000    | 0     | 959179080     |                   | 958560000  |               |          |
|                          | 10       | 200   | 738011460     | 729.73            | 743034380  | 734.82        | 0.99     |
|                          | 100      | 200   | 747347360     | 736.13            | 745493700  | 733.72        | 1.00     |
|                          | 1000     | 200   | 827614040     | 784.65            | 813552720  | 769.13        | 1.02     |
|                          | 2000     | 200   | 908938600     | 822.33            | 889483700  | 802.94        | 1.02     |
|                          | 4000     | 200   | 1100082180    | 915.35            | 1040240060 | 859.60        | 1.06     |
|                          | 8000     | 200   | 1489083780    | 1090.52           | 1373256720 | 974.42        | 1.12     |
|                          | 16000    | 200   | 2371317640    | 1412.14           | 2192029980 | 1233.47       | 1.14     |
|                          |          |       |               |                   |            |               |          |
| llama 8B Q4_K - Medium   |          |       |               |                   |            |               |          |
|                          | 10       | 0     | 11771200      |                   | 12052960   |               |          |
|                          | 100      | 0     | 17274700      |                   | 16183500   |               |          |
|                          | 1000     | 0     | 85221840      |                   | 85714320   |               |          |
|                          | 2000     | 0     | 175552220     |                   | 175143680  |               |          |
|                          | 4000     | 0     | 354803180     |                   | 354138120  |               |          |
|                          | 8000     | 0     | 758988000     |                   | 759214320  |               |          |
|                          | 16000    | 0     | 1724419420    |                   | 1723164020 |               |          |
|                          | 10       | 200   | 1371246100    | 1359.47           | 1333360260 | 1321.31       | 1.03     |
|                          | 100      | 200   | 1374515580    | 1357.24           | 1340339720 | 1324.16       | 1.02     |
|                          | 1000     | 200   | 1466855420    | 1381.63           | 1462192200 | 1376.48       | 1.00     |
|                          | 2000     | 200   | 1584856320    | 1409.30           | 1577855500 | 1402.71       | 1.00     |
|                          | 4000     | 200   | 1857644120    | 1502.84           | 1820829360 | 1466.69       | 1.02     |
|                          | 8000     | 200   | 2421499080    | 1662.51           | 2357915780 | 1598.70       | 1.04     |
|                          | 16000    | 200   | 3681208540    | 1956.79           | 3607611840 | 1884.45       | 1.04     |
|                          |          |       |               |                   |            |               |          |
| qwen2 1.5B Q4_K - Medium |          |       |               |                   |            |               |          |
|                          | 10       | 0     | 9674800       |                   | 7476440    |               |          |
|                          | 100      | 0     | 10231340      |                   | 10174900   |               |          |
|                          | 1000     | 0     | 31047760      |                   | 31299640   |               |          |
|                          | 2000     | 0     | 61244000      |                   | 61835680   |               |          |
|                          | 4000     | 0     | 125470380     |                   | 125735260  |               |          |
|                          | 8000     | 0     | 272910140     |                   | 272990700  |               |          |
|                          | 16000    | 0     | 654824360     |                   | 645511720  |               |          |
|                          | 10       | 200   | 581147620     | 571.47            | 561862220  | 554.39        | 1.03     |
|                          | 100      | 200   | 583839320     | 573.61            | 554249620  | 544.07        | 1.05     |
|                          | 1000     | 200   | 614533460     | 583.49            | 595459540  | 564.16        | 1.03     |
|                          | 2000     | 200   | 667034580     | 605.79            | 636267320  | 574.43        | 1.05     |
|                          | 4000     | 200   | 766202980     | 640.73            | 731192180  | 605.46        | 1.06     |
|                          | 8000     | 200   | 989000400     | 716.09            | 932145720  | 659.16        | 1.09     |
|                          | 16000    | 200   | 1508910500    | 854.09            | 1411803780 | 766.29        | 1.11     |
|                          |          |       |               |                   |            |               |          |
|                          |          |       |               |                   |            |               |          |
| qwen2 7B Q4_K - Medium   |          |       |               |                   |            |               |          |
|                          | 10       | 0     | 11509000      |                   | 10902840   |               |          |
|                          | 100      | 0     | 15888700      |                   | 15636600   |               |          |
|                          | 1000     | 0     | 83561980      |                   | 83783320   |               |          |
|                          | 2000     | 0     | 170899980     |                   | 170541560  |               |          |
|                          | 4000     | 0     | 346676380     |                   | 345789520  |               |          |
|                          | 8000     | 0     | 735418880     |                   | 734362580  |               |          |
|                          | 16000    | 0     | 1658389120    |                   | 1652635280 |               |          |
|                          | 10       | 200   | 1284247020    | 1272.74           | 1277166900 | 1266.26       | 1.01     |
|                          | 100      | 200   | 1284649740    | 1268.76           | 1285211980 | 1269.58       | 1.00     |
|                          | 1000     | 200   | 1390807060    | 1307.25           | 1386056060 | 1302.27       | 1.00     |
|                          | 2000     | 200   | 1519021700    | 1348.12           | 1493807840 | 1323.27       | 1.02     |
|                          | 4000     | 200   | 1783544660    | 1436.87           | 1714694020 | 1368.90       | 1.05     |
|                          | 8000     | 200   | 2322356100    | 1586.94           | 2202785980 | 1468.42       | 1.08     |
|                          | 16000    | 200   | 3547261460    | 1888.87           | 3321806240 | 1669.17       | 1.13     |

</details>

I have read the [contributing guidelines](https://github.com/ggerganov/llama.cpp/blob/master/CONTRIBUTING.md)
Self-reported review complexity:
- [x] Low
- [ ] Medium
- [ ] High