@gaugarg-nv 2025-03-17
Did some more testing on RTX 4090. I see around 10% improvement on Llama-3b and Qwen-1.5b model for high seq length.

  | ISL | OSL | Master: Gen phase tok/sec | PR: Gen phase tok/sec | Speed-up
-- | -- | -- | -- | -- | --
llama 3B Q4_K - Medium | 10 | 200 | 269.8302524 | 266.2338712 | 0.986672
  | 100 | 200 | 268.4936757 | 263.709292 | 0.982181
  | 1000 | 200 | 252.4435713 | 254.9205103 | 1.009812
  | 10000 | 200 | 170.5515393 | 189.4690656 | 1.11092
  |   |   |   |   |  
qwen2 1.5B   Q4_K - Medium |   |   |   |  
  | 10 | 200 | 344.4271769 | 360.4276128 | 1.046455
  | 100 | 200 | 342.1115136 | 360.8377528 | 1.054737
  | 1000 | 200 | 335.4593024 | 353.6292346 | 1.054164
  | 10000 | 200 | 264.7008325 | 288.7548641 | 1.090873