@gaugarg-nv 2025-03-19
Rebased the changes on top of master as there were some conflicts.

Also benchmarked performance on GTX 1080. This is what I get:

  | ISL | OSL | Master: Gen phase tok/sec | PR: Gen phase tok/sec | Speed-up
-- | -- | -- | -- | -- | --
llama 3B Q4_K - Medium | 100 | 100 | 64.40901405 | 66.44957568 | 1.031681
  | 1000 | 100 | 57.97212522 | 60.09357042 | 1.036594
  | 4000 | 100 | 48.94937194 | 49.50312464 | 1.011313
  |   |   |   |   |  
qwen2 1.5B   Q4_K - Medium |   |   |   |  
  | 100 | 100 | 82.36570957 | 84.50382656 | 1.025959
  | 1000 | 100 | 79.3190166 | 79.74659464 | 1.005391
  | 4000 | 100 | 68.77001876 | 75.35970213 | 1.095822

There are no regressions and we see a decent gain with Qwen 1.5B model. 

@JohannesGaessler do you think we can merge this change?