@ggerganov 2025-03-19
With V100 it seems there is not much difference - is this within expectations?

```bash
GGML_CUDA=1 ./scripts/compare-commits.sh master pr/12183 -m models-mnt/open-llama/7B-v2/ggml-model-q4_0.gguf -m models-mnt/open-llama/7B-v2/ggml-model-q8_0.gguf -m models-mnt/open-llama/7B-v2/ggml-model-q4_k.gguf -fa 1 -n 0 -p 4096 -ub 1,2,4,8,16,32,64,128,256,512 -r 1 -ngl 99
```

| Model           |   Microbatch size | Test   |   t/s master |   t/s pr/12183 |   Speedup |
|:----------------|------------------:|:-------|-------------:|---------------:|----------:|
| llama 7B Q4_0   |                 1 | pp4096 |       120.11 |         119.08 |      0.99 |
| llama 7B Q4_0   |                 2 | pp4096 |       207.93 |         211.72 |      1.02 |
| llama 7B Q4_0   |                 4 | pp4096 |       313.78 |         311.21 |      0.99 |
| llama 7B Q4_0   |                 8 | pp4096 |       487.66 |         493.68 |      1.01 |
| llama 7B Q4_0   |                16 | pp4096 |       756.47 |         763.94 |      1.01 |
| llama 7B Q4_0   |                32 | pp4096 |      1111.01 |        1114.67 |      1.00 |
| llama 7B Q4_0   |                64 | pp4096 |       613.40 |         607.49 |      0.99 |
| llama 7B Q4_0   |               128 | pp4096 |      1123.29 |        1102.10 |      0.98 |
| llama 7B Q4_0   |               256 | pp4096 |      1668.56 |        1672.35 |      1.00 |
| llama 7B Q4_0   |               512 | pp4096 |      2138.19 |        2157.27 |      1.01 |
| llama 7B Q4_K_M |                 1 | pp4096 |       112.75 |         112.06 |      0.99 |
| llama 7B Q4_K_M |                 2 | pp4096 |       177.46 |         177.55 |      1.00 |
| llama 7B Q4_K_M |                 4 | pp4096 |       258.02 |         259.64 |      1.01 |
| llama 7B Q4_K_M |                 8 | pp4096 |       310.70 |         312.40 |      1.01 |
| llama 7B Q4_K_M |                16 | pp4096 |       690.29 |         693.88 |      1.01 |
| llama 7B Q4_K_M |                32 | pp4096 |       998.44 |         995.80 |      1.00 |
| llama 7B Q4_K_M |                64 | pp4096 |       617.86 |         611.84 |      0.99 |
| llama 7B Q4_K_M |               128 | pp4096 |      1131.27 |        1111.56 |      0.98 |
| llama 7B Q4_K_M |               256 | pp4096 |      1679.40 |        1679.73 |      1.00 |
| llama 7B Q4_K_M |               512 | pp4096 |      2144.94 |        2161.94 |      1.01 |
| llama 7B Q8_0   |                 1 | pp4096 |        83.94 |          83.60 |      1.00 |
| llama 7B Q8_0   |                 2 | pp4096 |       148.76 |         150.36 |      1.01 |
| llama 7B Q8_0   |                 4 | pp4096 |       234.13 |         236.20 |      1.01 |
| llama 7B Q8_0   |                 8 | pp4096 |       382.40 |         385.34 |      1.01 |
| llama 7B Q8_0   |                16 | pp4096 |       642.34 |         644.91 |      1.00 |
| llama 7B Q8_0   |                32 | pp4096 |       947.97 |         949.16 |      1.00 |
| llama 7B Q8_0   |                64 | pp4096 |       817.02 |         809.33 |      0.99 |
| llama 7B Q8_0   |               128 | pp4096 |      1457.01 |        1426.50 |      0.98 |
| llama 7B Q8_0   |               256 | pp4096 |      2008.42 |        2011.89 |      1.00 |
| llama 7B Q8_0   |               512 | pp4096 |      2400.06 |        2426.38 |      1.01 |